requirements.txt
PyPDF2
pdf2image
pytesseract
reportlab
python-docx